+++ apps/stocks-api/src/environments/environment.prod.ts
@@ -1,3 +1,7 @@
-export const environment = {
-  production: true
+import { StocksAppConfig } from '@coding-challenge/stocks/data-access-app-config';
+
+export const environment: StocksAppConfig = {
+  production: true,
+  apiKey: '',
+  apiURL: 'https://cloud.iexapis.com'
 };
