+++ apps/stocks-api/src/environments/environment.ts
@@ -1,7 +1,11 @@
+import { StocksAppConfig } from '@coding-challenge/stocks/data-access-app-config';
+
 // This file can be replaced during build by using the `fileReplacements` array.
 // `ng build ---prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
-export const environment = {
-  production: false
+export const environment: StocksAppConfig = {
+  production: false,
+  apiKey: 'Tpk_c616780a1013445a8c3e65f21ec11053',
+  apiURL: 'https://sandbox.iexapis.com/'
 };
