+++ apps/stocks-api/src/server-config/server-config.ts
@@ -1,0 +1,7 @@
+/**
+ * Hapi server config
+ */
+export const server_config = {
+  port: 3333,
+  host: 'localhost'
+};
