+++ apps/stocks/src/environments/environment.prod.ts
@@ -1,7 +1,3 @@
-import { StocksAppConfig } from '@coding-challenge/stocks/data-access-app-config';
-
-export const environment: StocksAppConfig = {
-  production: true,
-  apiKey: '',
-  apiURL: 'https://sandbox.iexapis.com'
+export const environment = {
+  production: true
 };
